+++ jest.config.js
@@ -24,7 +24,9 @@ module.exports = {
     'jest-axe/extend-expect',
     '<rootDir>/jest.setup.js',
   ],
-  snapshotSerializers: ['jest-emotion'],
+  snapshotSerializers: [
+    '@emotion/jest/serializer'
+  ],
   testPathIgnorePatterns: [
     '/.cache/',
     '/coverage/',