+++ packages/charts/src/components/LineChart/LineChart.tsx
@@ -25,8 +25,7 @@ import { CustomScatterDataComponent } from './CustomPointComponent';
 import { CustomAxisComponent } from './CustomAxisComponent';
 import { LegendButton } from './LegendButton';
 
-import styled, { CreateStyled } from '@emotion/styled';
-const typedStyled = styled as CreateStyled<ThemeInterface>;
+import styled from '@emotion/styled';
 
 export type LineChartAxisStyle = VictoryAxisProps['style'];
 export type DataGetterPropType = VictoryLineProps['x'];
@@ -104,7 +103,7 @@ export interface LineChartProps<T extends ChartDataOptions> {
   y?: keyof T;
 }
 
-const LineChartContainer = typedStyled.div`
+const LineChartContainer = styled.div`
   max-height: 600px;
   max-width: 800px;
   svg {
@@ -112,13 +111,13 @@ const LineChartContainer = typedStyled.div`
   }
 `;
 
-const VictoryChartContainer = typedStyled.div``;
+const VictoryChartContainer = styled.div``;
 
-const DataLegendsContainer = typedStyled.div`
+const DataLegendsContainer = styled.div`
   padding-bottom: 24px;
 `;
 
-const DataLegendsDescription = typedStyled.p`
+const DataLegendsDescription = styled.p`
   color: ${(props: any) => props.theme.colors.neutral};
   font-size: ${(props: any) => props.theme.typeScale.size02.fontSize};
 `;