+++ packages/react-magma-dom/src/components/Drawer/Drawer.stories.tsx
@@ -30,6 +30,7 @@ export const Default = args => {
         header="Drawer Title"
         onClose={() => setShowDrawer(false)}
         isOpen={showDrawer}
+        closeAriaLabel="Close drawer"
         {...args}
       >
         <p>This is a Drawer, doing Drawer things.</p>
@@ -54,6 +55,8 @@ export const SiteNavigation = args => {
         onClose={() => setShowDrawer(false)}
         isOpen={showDrawer}
         position={DrawerPosition.right}
+        ariaLabel="Site Navigation Drawer"
+        closeAriaLabel="Close Navigation Drawer"
       >
         <NavTabs orientation={TabsOrientation.vertical}>
           <NavTab to="#">One</NavTab>