+++ packages/react-magma-dom/src/components/Heading/index.tsx
@@ -62,7 +62,7 @@ export const Heading = React.forwardRef<HTMLHeadingElement, HeadingProps>(
     return (
       <TypographyComponent
         {...other}
-        as={headingElement}
+        as={headingElement as any}
         contextVariant={contextVariant}
         data-testid={testId}
         isInverse={useIsInverse(props.isInverse)}