+++ packages/react-magma-dom/src/components/TreeView/useTreeView.ts
@@ -59,6 +59,19 @@ export interface UseTreeViewProps {
   onSelectedItemChange?: (
     selectedItems: Array<TreeItemSelectedInterface>
   ) => void;
+  /**
+   * Only affects if selectable mode is TreeViewSelectable.multi.
+   * Determines if the parent checkbox will get selected when the user selects all its children checkboxes.
+   * When checkParents is enabled, the TreeView displays the indeterminate state of the parent checkboxes too.
+   * @default true
+   */
+  checkParents?: boolean;
+  /**
+   * Only affects if selectable mode is TreeViewSelectable.multi.
+   * Determines if the child checkboxes get selected when the user selects parent checkbox.
+   * @default true
+   */
+  checkChildren?: boolean;
   children?: React.ReactNode[];
 }
 
@@ -69,6 +82,8 @@ export function useTreeView(props: UseTreeViewProps) {
     onExpandedChange,
     initialExpandedItems,
     preselectedItems,
+    checkChildren = true,
+    checkParents = true,
     children,
   } = props;
 
@@ -155,6 +170,8 @@ export function useTreeView(props: UseTreeViewProps) {
     setInitialExpandedItemsNeedUpdate,
     selectedItemsChanged,
     setSelectedItemsChanged,
+    checkChildren,
+    checkParents
   };
 
   return { contextValue };