+++ packages/charts/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @react-magma/charts
 
+## 9.0.0-next.0
+
+### Patch Changes
+
+- Updated dependencies [fdf2dc929]
+  - react-magma-dom@4.6.0-next.0
+
+## 8.0.0-next.5
+
+### Minor Changes
+
+- ae668a3e5: Updating emotion to v11
+
+## 8.0.0-next.4
+
+### Patch Changes
+
+- Updated dependencies [9e38e9e7f]
+  - react-magma-dom@4.5.0-next.7
+
+## 7.0.1-next.3
+
+### Patch Changes
+
+- 9f43ba1e6: fix(CarbonCharts): Minify styles
+
 ## 7.0.1-next.2
 
 ### Patch Changes